Move runEnvironment into its own file
